requirements.txt
google-generativeai
python-dotenv
Jinja2